+++ WebAPI/.dockerignore
