+++ WebAPI/WebAPI.csproj
@@ -12,4 +12,8 @@
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.4.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Application\Interactors\IsFeatureFlagEnabled\" />
+  </ItemGroup>
+
 </Project>
